+++ src/rpcserver.h
@@ -78,7 +78,7 @@ extern void ShutdownRPCMining();
 
 extern int64_t nWalletUnlockTime;
 extern int64_t AmountFromValue(const json_spirit::Value& value);
-extern json_spirit::Value ValueFromAmount(int64_t amount);
+extern json_spirit::Value ValueFromAmount(__int128 amount);
 extern double GetDifficulty(const CBlockIndex* blockindex = NULL);
 
 extern double GetPoWMHashPS();
